move FilesSelector to its component
