+++ catalog/app/containers/Bucket/PackageDialog/index.ts
@@ -2,6 +2,7 @@ export { default as DialogError } from './DialogError'
 export { default as DialogLoading } from './DialogLoading'
 export { default as DialogSuccess } from './DialogSuccess'
 export * from './FilesInput'
+export * from './FilesSelector'
 export * from './Layout'
 export * from './Manifest'
 export * from './MetaInput'
